feat: chargement instantané des 50 1ers Pokémons
Cette méthode peut être clairement améliorée, via initialData passé au useInfiniteQuery et potentiellement aussi au trpc.useQueries
