+++ src/app/components/PokemonList/PokemonCard.tsx
@@ -1,0 +1,30 @@
+import Link from "next/link";
+import { Pokemon } from "~/libs/poke/dto/pokemon";
+
+interface PokemonCardProps {
+  pokemon: Pokemon;
+}
+
+export function PokemonCard({ pokemon }: PokemonCardProps) {
+  return (
+    <div
+      className="flex flex-col justify-center items-center rounded-md border border-gray-600 bg-gray-700 p-4 w-[150px] h-[200px]"
+    >
+      <h2 className="max-w-[100%] text-[20px] font-bold capitalize truncate">
+        {pokemon?.name}
+      </h2>
+      {pokemon?.sprites.front_default && (
+        <img
+          src={pokemon?.sprites.front_default ?? ''}
+          className="max-w-[80px]"
+        />
+      )}
+      <Link
+        href={`/pokemon/${pokemon?.id}`}
+        className="mt-2 inline-block text-center rounded-md bg-indigo-500 px-4 py-2 text-white hover:bg-indigo-600"
+      >
+        View
+      </Link>
+    </div>
+  )
+}
